New branch4

+++ backend/api.py
@@ -1,5 +1,7 @@
 
+from tokenize import Double, String
 from typing import List
+from unicodedata import decimal
 from ninja import Router
 from backend.models import Author, Book, Genre, Items, Saved_Books
 from backend.schemas import AuthorsSchemaOut, BookSchemaIn, BookSchemaOut, ErrorMesssage, GenresSchemaOut, ItemsSchemaIn, ItemsSchemaOut, SavedBookSchemaOut
@@ -34,7 +36,7 @@ def myFunction5(request, author_name: str):
     try:
         return Author.objects.get(name=author_name).books.all()
     except Exception:
-        return 400,  {"detial": "Author Doesnt Exists"}
+        return 400,  {"detail": "Author Doesnt Exists"}
 
 
 @myRouters.get("/get_all_books_by_genre/{genre_name}", response={200: List[BookSchemaOut], 400: ErrorMesssage})
@@ -42,7 +44,7 @@ def myFunction6(request, genre_name: str):
     try:    
         return Genre.objects.get(name=genre_name).books.all()
     except Exception:
-        return 400,  {"detial": "Genre Doesnt Exists"}
+        return 400,  {"detail": "Genre Doesnt Exists"}
 
 
 #when u press save or unsave button
@@ -56,9 +58,11 @@ def myFunction7(request, user_ids: int, book_ids: int, saveCondition: bool):
             return {"msg": "Book already saved"}
         Saved_Books.objects.create(user_id=user_ids, book_id=book_ids, saved=True)
         return {"msg": "saved successfully"}
-    objectss = Saved_Books.objects.filter(user_id=user_ids, book_id=book_ids)
-    objectss.delete()
-    return {"msg": "unsaved successfully"}
+    if Saved_Books.objects.filter(user_id=user_ids, book_id= book_ids).exists() and not saveCondition:
+        objectss = Saved_Books.objects.filter(user_id=user_ids, book_id=book_ids)
+        objectss.delete()
+        return {"msg": "unsaved successfully"}
+    return {"msg": "Book Not Saved Yet"}
 
 
 #when u press on saved
@@ -70,7 +74,7 @@ def myFunction8(request, user_ids: int):
 # when u press on add to cart button or remove from cart
 @myRouters.post("/add_remove_cart_items")
 def myFunction9(request, desiredBook: ItemsSchemaIn):
-    if(desiredBook.removeFromCart == False):
+    if (not(Items.objects.filter(user_id=desiredBook.user_id,book_id=desiredBook.book_id).exists())and  desiredBook.removeFromCart == False):
         Items.objects.create(
             user_id=desiredBook.user_id,
             book_id=desiredBook.book_id,
@@ -78,17 +82,36 @@ def myFunction9(request, desiredBook: ItemsSchemaIn):
             inCart=True
         )
         return {"msg": "Book Added Successfully",}
-    objectss = Items.objects.filter(user_id = desiredBook.user_id, book_id = desiredBook.book_id, qty = desiredBook.qty)
-    objectss.delete()
-    return {"msg": "Book Removed Successfully",}
+
+    elif (Items.objects.filter(user_id=desiredBook.user_id,book_id=desiredBook.book_id, inCart = False).exists() and desiredBook.removeFromCart == False):
+        objects = Items.objects.get(user_id=desiredBook.user_id,book_id=desiredBook.book_id,inCart=False,)
+        objects.qty = 0
+        objects.save()
+        objects.qty += desiredBook.qty
+        objects.inCart = True
+        objects.save()
+        return {"msg": "You Already Checked This Book But Its Added Anyway!",}
+
+    elif (Items.objects.filter(user_id=desiredBook.user_id,book_id=desiredBook.book_id, inCart = True).exists() and desiredBook.removeFromCart == False):
+        objects = Items.objects.get(user_id=desiredBook.user_id,book_id=desiredBook.book_id,inCart=True)
+        objects.qty += desiredBook.qty
+        objects.save()
+        return {"msg": "Qty Increased Successfully",}
+
+    elif(Items.objects.filter(user_id=desiredBook.user_id,book_id=desiredBook.book_id,inCart=True).exists()):
+        objectss = Items.objects.get(user_id = desiredBook.user_id, book_id = desiredBook.book_id)
+        objectss.inCart = False
+        #objectss.isBought = True
+        objectss.save()
+        return {"msg": "Book Removed Successfully",}
 
 
 # when u press on the cart button
-@myRouters.get("/get_personal_cart/{user_ids}", response={200:List[ItemsSchemaOut], 400: ErrorMesssage})
+@myRouters.get("/get_personal_cart_items/{user_ids}", response={200:List[ItemsSchemaOut], 201: ErrorMesssage})
 def myFunction10(request, user_ids: int):
     if data:= Items.objects.filter(user_id=user_ids, inCart=True):
         return data
-    return 400, {"detial": "Your Cart Is Empty"}
+    return 201, {"detail": "Your Cart Is Empty"}
 
 
 # when u press on the cart button... and you should update it when remove item from cart 
@@ -102,7 +125,7 @@ def myFunction12(request, user_ids: int):
         qty = datum_['qty']
         objects = Book.objects.get(id=datum_['book_id'])
         totalqty += qty
-        totalPrice += qty * objects.price
+        totalPrice += ((qty) * (objects.price))
     return {"totalPrice": totalPrice, "totalqty": totalqty}
 
 
@@ -112,15 +135,17 @@ def myFunction13(request, user_ids: int):
     booksInCart = Items.objects.filter(user_id=user_ids, inCart=True)
     if data := list(booksInCart.values()):
         try:
-            for i in range(len(data)):
-                qty = data[i]['qty']
-                objects = Book.objects.get(id=data[i]['book_id'])
-                objects.total_sales += qty
-                objects.save()
-                objects = booksInCart[i]
-                objects.inCart = False
-                objects.isBought = True
-                objects.save()
+            for datum in data:
+                qty = datum['qty']
+                bookObjects = Book.objects.get(id=datum['book_id'])
+                bookObjects.total_sales += qty
+                bookObjects.save()
+                itemObjects = booksInCart[0]
+                print(booksInCart)
+                itemObjects.qty = 0
+                itemObjects.inCart = False
+                itemObjects.isBought = True
+                itemObjects.save()
             return {"msg": "Items Bought successfully"}
         except Exception:
             return {"msg": "Try Again Please"}
@@ -128,12 +153,12 @@ def myFunction13(request, user_ids: int):
 
 
 # when u press on your profile
-@myRouters.get("/get_all_purchased_books/{user_ids}", response=List[ItemsSchemaOut])
+@myRouters.get("/get_all_purchased_books/{user_ids}", response={200:List[ItemsSchemaOut], 201: ErrorMesssage})
 def myFunction14(request, user_ids: int):
     try:
         return Items.objects.filter(user_id=user_ids, isBought=True)
     except Exception:
-        return{"msg": "No Purchased Books"}
+        return 201, {"detail": "No Purchased Books"}
 
 
 # Top sales
